+++ Source/Experiments/Ncd/ThresholdTask/NcdThresholdTask.m
@@ -153,7 +153,7 @@
     return width;
 }
 
-- (void)setWidth:(float)aWidth 
+- (void)setWidth:(CGFloat)aWidth 
 {
     [[[self undoManager] prepareWithInvocationTarget:self] setWidth:width];
     width = aWidth;
