+++ Source/Main Dialogs/CommandCenter/ORCommandCenterController.m
@@ -213,14 +213,14 @@ SYNTHESIZE_SINGLETON_FOR_ORCLASS(CommandCenterController);
 
 
 #pragma mark •••Data Source Methods
-- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(int) rowIndex
+- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(NSInteger) rowIndex
 {
     id obj = [[[self commandCenter] clients]  objectAtIndex:rowIndex];
     return [obj valueForKey:[aTableColumn identifier]];
 }
 
 // just returns the number of items we have.
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     return [[[self commandCenter] clients] count];
 }
