+++ Source/Objects/Control/RunList/ORRunListController.m
@@ -376,7 +376,7 @@
 }
 
 #pragma mark Data Source Methods
-- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(int) rowIndex
+- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(NSInteger) rowIndex
 {
 	if(aTableView == itemsListView){
         if((rowIndex == 0) && [[aTableColumn identifier] isEqualToString:@"SubRun"])return nil;
@@ -387,12 +387,12 @@
 	else return nil;
 }
 
-- (void)tableView:(NSTableView *)aTableView willDisplayCell:(id)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex
+- (void)tableView:(NSTableView *)aTableView willDisplayCell:(id)aCell forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex
 {
     if(rowIndex == 0 && [[aTableColumn identifier] isEqualToString:@"SubRun"])[aCell setEnabled:NO];
     else [aCell setEnabled:YES];
 }
-- (void) tableView:(NSTableView *)aTableView setObjectValue:(id)anObject forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex
+- (void) tableView:(NSTableView *)aTableView setObjectValue:(id)anObject forTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex
 {
 	if(aTableView == itemsListView){
         if((rowIndex == 0) && [[aTableColumn identifier] isEqualToString:@"SubRun"])return;
@@ -402,7 +402,7 @@
 }
 
 // just returns the number of items we have.
-- (int) numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger) numberOfRowsInTableView:(NSTableView *)aTableView
 {
 	if(aTableView == itemsListView){
 		return [model itemCount];
